% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\alias{n_nodes}
\alias{n_edges}
\title{Nodes and edges}
\usage{
n_nodes(n3d, nodes, ...)

n_edges(n3d, edges, ...)
}
\arguments{
\item{n3d}{An object of class \code{n3d} as returned by \code{\link{n3d}}.}

\item{nodes, edges}{Nodes and edges data.frames to add to the network.}

\item{...}{Any column(s) from the \code{nodes} or \code{edges} to select, can be named.}
}
\description{
Set nodes and edges of the network
}
\examples{
n3d() \%>\% 
  n_nodes(lesmis_nodes, id, name, color, value = out_degree) \%>\% 
  n_edges(lesmis_edges, source, target)

}
