% tinyrox says don't edit this manually, but it can't stop you!
\name{n3d_revision}
\alias{n3d_revision}
\title{Model Revision}
\usage{
n3d_revision()
}
\value{
A 40-character commit hash.
}
\description{
The Hugging Face commit of \code{nvidia/Nemotron-3-Diarization} that
\code{\link{load_n3d}} and \code{\link{download_n3d}} use by default.
}
\examples{
n3d_revision()
}
