% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force.R
\name{engine}
\alias{engine}
\alias{n_force_engine}
\title{Force Engineß}
\usage{
n_force_engine(n3d, engine = c("d3", "ngraph"))
}
\arguments{
\item{n3d}{An object of class \code{n3d} as returned by \code{\link{n3d}}.}

\item{engine}{Getter/setter for which force-simulation engine to use 
(\href{https://github.com/vasturiano/d3-force-3d}{d3} or \href{https://github.com/anvaka/ngraph.forcelayout}{ngraph}).}
}
\description{
Define the force engine to use.
}
